+++ common/dircmp.c
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2000-2005 MAEKAWA Masahide <maekawa@cvsync.org>
+ * Copyright (c) 2000-2013 MAEKAWA Masahide <maekawa@cvsync.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -308,8 +308,8 @@ dircmp_start(struct dircmp_args *dca, const char *name, const char *relname)
 
 	SetWord(cmd, len - 2);
 	cmd[2] = FILESCAN_START;
-	cmd[3] = namelen;
-	cmd[4] = relnamelen;
+	cmd[3] = (uint8_t)namelen;
+	cmd[4] = (uint8_t)relnamelen;
 	if (!mux_send(dca->dca_mux, MUX_FILESCAN, cmd, 5))
 		return (false);
 	if (!mux_send(dca->dca_mux, MUX_FILESCAN, name, namelen))
