+++ common/filecmp.c
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2000-2005 MAEKAWA Masahide <maekawa@cvsync.org>
+ * Copyright (c) 2000-2013 MAEKAWA Masahide <maekawa@cvsync.org>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -284,8 +284,8 @@ filecmp_start(struct filecmp_args *fca, const char *name, const char *relname)
 
 	SetWord(cmd, len - 2);
 	cmd[2] = UPDATER_START;
-	cmd[3] = namelen;
-	cmd[4] = relnamelen;
+	cmd[3] = (uint8_t)namelen;
+	cmd[4] = (uint8_t)relnamelen;
 	if (!mux_send(fca->fca_mux, MUX_UPDATER, cmd, 5))
 		return (false);
 	if (!mux_send(fca->fca_mux, MUX_UPDATER, name, namelen))
